+++ packages/core/client.d.ts
@@ -72,4 +72,6 @@ export default class ClientWithInternals<T extends Config = Config> extends Clie
   }
 
   _loadPlugin(plugin: Plugin): void
+
+  _isBreadcrumbTypeEnabled(type: string): boolean
 }