+++ packages/plugin-react-native-app-state-breadcrumbs/app-state.js
@@ -2,7 +2,7 @@ const { AppState } = require('react-native')
 
 module.exports = {
   load: client => {
-    if (!client._config.enabledBreadcrumbTypes || !client._config.enabledBreadcrumbTypes.includes('state')) return
+    if (!client._isBreadcrumbTypeEnabled('state')) return
 
     AppState.addEventListener('change', state => {
       client.leaveBreadcrumb('App state changed', { state }, 'state')