+++ test/node/features/steps/server_fixture_request_steps.rb
@@ -7,3 +7,12 @@ require 'net/http'
 When("I POST the data {string} to the URL {string}") do |reqbody, url|
   Net::HTTP.post(URI(url), reqbody)
 end
+
+
+When('I open the URL {string} tolerating any error') do |url|
+  begin
+    open(url, &:read)
+  rescue
+    $logger.debug $!.inspect
+  end
+end